% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPredictorCorrelatedFeatures.R
\name{addPredictorCorrelatedFeatures}
\alias{addPredictorCorrelatedFeatures}
\title{Add features that are correlated with a set of elastic net predictors.}
\usage{
addPredictorCorrelatedFeatures(enResults, dataMatList, corThreshold = 0.8)
}
\arguments{
\item{enResults}{an elastic net results object.}

\item{dataMatList}{a list of matrices with feature data organized along the rows,
and feature names accessible via rownames(dataMatList).  These matrices must contain
data for all features used in the original elastic net call.}

\item{corThreshold}{a correlation threshold.}
}
\value{
an elastic net results object containing all data in the enResults parameter,
plus two added elements:
\itemize{
  \item{predictorCorFeatures} {is a list of character vectors indexed by predictor name.  
Each character vector associated with a predictor contains the names of all features correlated 
above corThreshold with that predictor. (Correlations are with respect to features used as 
input to the original elastic net call, as recorded in rownames(enResults$featureWtMat).)}
  \item{predictorsAndCorFeatures} {is a character vector containing all predictors together 
with features correlated with them above corThreshold.}
}
}
\description{
Add features that are correlated with a set of elastic net predictors.
}
\concept{
rcellminerElasticNet
}
