% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvElasticNet.R
\name{cvElasticNet}
\alias{cvElasticNet}
\title{Cross validate the entire elastic net procedure to obtain a set of predicted
response values for strictly held out data.}
\usage{
cvElasticNet(nFolds = 10, nRepeats = 10, cvFoldIds = NULL,
  minFeatureResponseAbsCor = 0, featureMat, responseVec, standardize = TRUE,
  alphaVals = seq(0.2, 1, length = 9), lambdaVals = NULL,
  nFoldsForParamSelection = 10, nCvRepeats = 10, nTrainingRuns = 200,
  minFeatureFrequencyPctl = 0.95, cumCorNumPredictors = 10,
  useLambdaMin = TRUE, useOneStdErrRule = FALSE, id = "",
  keepEnResults = FALSE, verbose = TRUE)
}
\arguments{
\item{nFolds}{The number of folds used in cross-validation.}

\item{nRepeats}{The number of cross-validation repeats (with different fold
assignments).}

\item{cvFoldIds}{An nObs x nReplicates matrix with integer fold identifiers 
along the columns.}

\item{minFeatureResponseAbsCor}{Minimum absolute correlation between feature
and response (used to filter rows of featureMat).}

\item{featureMat}{p x n matrix with input feature vectors along rows.}

\item{responseVec}{n-dimensional response vector to be predicted using a sparse
linear combination of input feature vectors specified in featureMat.}

\item{standardize}{Logical flag for feature variable standardization (across
observations), passed in to glmnet functions (glmnet documentation: if
variables are already in the same units, standardization may not be necessary).}

\item{alphaVals}{a vector of alpha values to be optimized over}

\item{lambdaVals}{a vector of lambda values to be optimized over}

\item{nFoldsForParamSelection}{the number of cross-validation folds to perform}

\item{nTrainingRuns}{number of training runs to perform}

\item{minFeatureFrequencyPctl}{a fractional value (0-1). Features in the x 
percentile defined by this parameter after the removal of features with 
zero weights are retained.}

\item{cumCorNumPredictors}{the maximum number of predictors to be returned}

\item{useLambdaMin}{a boolean, whether to use the minimum lambda (lambda.min) 
from cross-validation as the optimum lambda, if FALSE then the largest value 
of lambda such that error is within 1 standard error of the minimum (lambda.1se)
is used.}

\item{useOneStdErrRule}{Use one standard error rule for model selection, i.e.,
select smallest model (in terms of number of predictors) for which the estimated
error (by cross-validation) is within one standard error of the minimum estimated
error.}

\item{id}{a optional string identifier for the EN run}

\item{verbose}{a boolean, whether debugging information should be displayed}
}
\value{
A list with the following elements
\itemize{
\item{cvPred} {A vector of cross-validation predicted values (averaged over cross validation
 repeats).}
\item{cvPredRsqared} {The square of the Pearson's correlation between the above vector of
 predicted response values and the actual response values.}
\item{cvMeanSqErr} {The mean squared error (averaged over the results from nFolds x nRepeats
 sets of cross-validation predictions).}
\item{cvSdMeanSqErr} {The standard deviation of the above (nFolds x nRepeats) cross validation
 mean squared error values.}
\item{cvPredMat} {An nObservations x nRepeats matrix with entry i, j indicating 
 cross-validation predicted response for the i-th observation in the j-th
 cross-validation repeat.}
\item{cvMeanSqErrMat} {An nFolds x nRepeats matrix with entry i, j indicating the 
 the mean cross validation error in the i-th fold of the j-th cross-validation
 repeat.}
\item{cvFoldEnPredictorWts} {A feature x en-run matrix, with cells indicating the weight 
 associated with a particular feature in a particular elastic net run.  (A zero 
 entry indicates that a given feature was not selected in the run associated with 
 the matrix column.)}
\item{cvFoldEnPredictorTab} {A data frame with summary information for predictors selected
 over cv-fold elastic net runs (selection frequency, average weight, sign of 
 predictor weights).}
}
}
\description{
Cross validate the entire elastic net procedure to obtain a set of predicted
response values for strictly held out data.
}
\concept{
rcellminerElasticNet
}

