% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCvFoldIds.R
\name{getCvFoldIds}
\alias{getCvFoldIds}
\title{Return a matrix of fold assignments for use in cross-validation.}
\usage{
getCvFoldIds(nObs, nFolds = 10, nRepeats = 10)
}
\arguments{
\item{nObs}{The number of observations in the data set.}

\item{nFolds}{The number of folds used in cross-validation.}

\item{nRepeats}{The number of cross-validation repeats (with different fold
assignments).}

\item{id}{a optional string identifier for the EN run}
}
\value{
An nObs x nReplicates matrix with integer fold identifiers along
 the columns.
}
\description{
Return a matrix of fold assignments for use in cross-validation.
}
\concept{
rcellminerElasticNet
}
