% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFdaClinicalTrialNscs.R
\name{getFdaClinicalTrialNscs}
\alias{getFdaClinicalTrialNscs}
\title{Get a list of NSCs that are FDA approved on in clinical trials}
\usage{
getFdaClinicalTrialNscs(file = NULL)
}
\arguments{
\item{file}{a file name, if provided a file will be generated at this location with the results}
}
\value{
a vector of NSCs annotated with an FDA status of FDA approved or 
  in clinical trials for rcellminer data
}
\description{
Get a list of NSCs that are FDA approved on in clinical trials
}
\examples{
getFdaClinicalTrialNscs()

}
\concept{rcellminerElasticNet}
