% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLmEquationString.R
\name{getLmEquationString}
\alias{getLmEquationString}
\title{Generate an equation string representing the model}
\usage{
getLmEquationString(predictorWts, orderByDecrAbsVal = TRUE, numSigDigits = 3)
}
\arguments{
\item{predictorWts}{a named vector of numeric weights}

\item{orderByDecrAbsVal}{boolean, whether to order predictors or keep the input order}

\item{numSigDigits}{number of digits to show}
}
\value{
a string representing the model
}
\description{
Generate an equation string representing the model
}
\author{
Vinodh Rajapakse
}
\concept{rcellminerElasticNet}
