% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiTaskElasticNet.R
\name{multiTaskElasticNet}
\alias{multiTaskElasticNet}
\title{The multi-response elastic net (EN) function with multiple training run averaging.}
\usage{
multiTaskElasticNet(
  X,
  Y,
  alphaVals = seq(0.2, 1, length = 9),
  lambdaVals = NULL,
  nFoldsForParamSelection = 4,
  nTrainingRuns = 200,
  minFeatureFrequencyPctl = 0.95,
  verbose = TRUE,
  useLambda1se = TRUE,
  id = ""
)
}
\arguments{
\item{X}{n x p matrix with input feature vectors along columns.}

\item{Y}{n x k matrix of responses to be predicted using sparse
linear combinations of input feature vectors specified in X}

\item{alphaVals}{a vector of alpha values to be optimized over}

\item{lambdaVals}{a vector of lambda values to be optimized over}

\item{nFoldsForParamSelection}{the number of cross-validation folds to perform}

\item{nTrainingRuns}{number of training runs to perform}

\item{minFeatureFrequencyPctl}{a fractional value (0-1). Features in the x 
percentile defined by this parameter after the removal of features with 
zero weights are retained.}

\item{verbose}{a boolean, whether debugging information should be displayed}

\item{useLambda1se}{a boolean, whether to use the largest value of lambda such that the error 
is within 1 standard error of the minimum (lambda.1se) as the optimum lambda;
if FALSE then lambda.min is used.}

\item{id}{a optional string identifier for the EN run}
}
\value{
a list of enResults objects, each with members: 
\itemize{
  \item{predictorWts} {weights for selected predictors}
  \item{predictorSelectionFreq} {the frequency of the selected predictors across EN iterations}
  \item{predictedResponse} {cumulative predicted response across the predictors}
  \item{predictedResponseCor} {a vector whose nth element gives the correlation of 
    the response with the linear combination of the first n predictors, with 
    coefficient weights given in predictorWts.}
  \item{call} {the command used to call the function with parameters described}
  \item{alpha} {the optimized alpha value used}
  \item{lambda} {the optimized lambda value used}
  \item{foldIds} {the cross-validation IDs used}
  \item{cvm} {mean cross-validated errors for alpha values tried}
  \item{cvOutput} {an object containing all cross-validation information for the optimized alpha}
  \item{featureWtMat} {a matrix with the feature weights for all features across all training runs}
}
}
\description{
Applies the elastic net regression algorithm to learn a sparse linear model
for predicting a response matrix from a set of input feature vectors.
}
\author{
Vinodh Rajapakse
}
\concept{rcellminerElasticNet}
