% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotElasticNet.R
\name{plotElasticNet}
\alias{plotElasticNet}
\title{Plot elastic net results}
\usage{
plotElasticNet(drugName, weights, drugAct, molDb, numCellLines,
  numFeatures = 10, featureAnnotations = NULL, showCellLineLabels = TRUE,
  pdfFilename = NULL, thresholdValues = FALSE, isMutBin = TRUE,
  axisScaling = NULL, topMarWt = 4, bottomMarWt = 2.75,
  leftMarHeatmap = 5, pdfWidth = 10, responseLabel = "Drug Activity",
  verbose = TRUE)
}
\arguments{
\item{drugName}{a string with the name of drug}

\item{weights}{a named vector of values with the elastic net weights for a set of features}

\item{drugAct}{a matrix of drug activity values, a value for each drug tested for each cell line}

\item{molDb}{a list of matrix of supported data types. The list must use the following prefixes: 
exp: gene expression, mir: miRNA, mut: mutations, pro: protein, mda: metadata, 
cop: copy number variations}

\item{numCellLines}{an integer of the number of cell lines to be plotted. This function will take the top (by drug activity)
numCellLines/2 and the bottom numCellLines/2 for plotting}

\item{numFeatures}{an integer of the number of features to be plotted; default 10}

\item{featureAnnotations}{a vector of optional annotations that will be appended to feature names in parentheses}

\item{showCellLineLabels}{a boolean whether to show cell line labels; default TRUE}

\item{pdfFilename}{a string filename for saving the image to a PDF}

\item{thresholdValues}{a boolean to threshold values to an upper limit of 3 and a lower bound of -3}

\item{isMutBin}{a boolean whether the mutation values are binarized (default: TRUE)}

\item{axisScaling}{a scaling factor for the the feature labels, if not use a 
default will be used depending on if the there annotations}

\item{topMarWt}{a user selected top margin for the the feature weights (default: 4)}

\item{bottomMarWt}{a user selected top margin for the the feature weights (default: 2.75)}

\item{leftMarHeatmap}{the left margin for the the heatmap and drug activity; 
can be increased to handle longer labels (default: 5)}

\item{pdfWidth}{the width of the PDF (default: 10)}

\item{responseLabel}{a string describing the response vector (default: "Drug Activity")}

\item{verbose}{a boolean, whether to display debugging information}
}
\value{
if verbose is TRUE then a list with the plotted drug activity values, heatmap values, and 
  cell line order will be returned

plotElasticNet does not return values; a plot is produced
}
\description{
Creates a plot for elastic net results similar to the GDSC plots (http://www.cancerrxgene.org/help/)
}
\details{
This function uses layout() and may conflict if embedded in other plots using layout().
  If the pdfFilename argument is used then the image will not be displayed on the screen.
}
\note{
Make sure the numFeatures is less than the length of the weights
}
\examples{
#plotElasticNet("PD-0325901", weights, binMutCCLE, drugActCCLE, genExpCCLE, 400, 10)

}
\references{
GDSC Website: http://www.cancerrxgene.org/help/
}
\author{
Augustin Luna
}
\concept{
rcellminerElasticNet
}
