% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectElasticNetParams.R
\name{selectElasticNetParams}
\alias{selectElasticNetParams}
\title{Select the optimal elastic net parameters based on cross-validation
error estimates.}
\usage{
selectElasticNetParams(featureMat, responseVec, standardize = TRUE,
  standardizeY = FALSE, fitIntercept = TRUE, alphaVals = seq(0.2, 1,
  length = 9), lambdaVals = NULL, nFolds = 10, nRepeats = 10,
  useLambdaMin = TRUE, verbose = TRUE)
}
\arguments{
\item{featureMat}{p x n matrix with input feature vectors along rows.}

\item{responseVec}{n-dimensional response vector to be predicted using a sparse
linear combination of input feature vectors specified in featureMat.}

\item{standardize}{Logical flag for feature variable standardization (across
observations), passed in to glmnet functions (glmnet documentation: if
variables are already in the same units, standardization may not be necessary).}

\item{standardizeY}{Logical flag for response variable standardization across
observations.}

\item{fitIntercept}{Logical flag indicating whether intercept term should be fit.}

\item{alphaVals}{a vector of alpha values to be optimized over}

\item{lambdaVals}{a vector of lambda values to be optimized over}

\item{nFolds}{The number of folds used in cross-validation.}

\item{nRepeats}{The number of cross-validation repeats (with different fold
assignments).}

\item{useLambdaMin}{a boolean, whether to use the minimum lambda (lambda.min) 
from cross-validation as the optimum lambda, if FALSE then the largest value 
of lambda such that error is within 1 standard error of the minimum (lambda.1se)
is used.}
}
\value{
A list with the following elements:
\itemize{
 \item{alpha} {The optimal alpha parameter.}
 \item{lambda} {The optimal lambda parameter.}
 \item{minCvError} {The minimum average cross validation error.}
 \item{cvInfoTab} {A data frame summarizing minimum cross validation errors and lambda
  parameter selections for each possible choice of the alpha parameter.}
 \item{cvFoldIds} {A matrix indicating the fold ids for each cross validation repeat
  along the columns.}
}
}
\description{
Select the optimal elastic net parameters based on cross-validation
error estimates.
}
\concept{
rcellminerElasticNet
}
