% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.enResults.R
\name{summary.enResults}
\alias{summary.enResults}
\title{Summary of Elastic Net Results}
\usage{
\method{summary}{enResults}(object, ...)
}
\arguments{
\item{object}{an object of type enResults}

\item{...}{additional parameters}
}
\value{
Nothing
}
\description{
Summary of Elastic Net Results
}
\concept{rcellminerElasticNet}
